LPD-59060 Inline BooleanFilterTranslator/BooleanFilterTranslatorImpl
